move checkpoint helpers
